#17 Rename tyre temps component
