+++ src/dashboard/src/components/charts/TyreTempsWidget.vue
